move clone_on_copy to its own module
